+++ ydb/tools/mnc/cli/commands/install.py
@@ -5,9 +5,8 @@ import os.path
 import rich
 
 
-from ydb.tools.mnc.lib import agent_client, common, configs, deploy, deploy_ctx, init, progress, service, tools
+from ydb.tools.mnc.lib import agent_client, common, configs, deploy, deploy_ctx, init, progress, service, term, tools
 from ydb.tools.mnc.scheme import multinode
-from ydb.tools.mnc.lib.draft import term
 
 from . import disks, uninstall
